requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
aiofiles==23.2.1
jinja2==3.1.3
google-auth==2.27.0
google-auth-oauthlib==1.2.0
python-jose[cryptography]==3.3.0
pyjwt==2.8.0 